+++ cranko-release-info-v1
[[projects]]
qnames = ["WWTCoreUWP", "csproj"]
version = "6.0.2.0"
age = 8

[[projects]]
qnames = ["UwpRenderEngine", "csproj"]
version = "6.0.2.0"
age = 8

[[projects]]
qnames = ["WWTHolographic", "csproj"]
version = "6.0.2.0"
age = 8

[[projects]]
qnames = ["OculusWrap", "csproj"]
version = "2.3.5.0"
age = 9

[[projects]]
qnames = ["MakeDataCabinetFile", "csproj"]
version = "1.0.1.0"
age = 8

[[projects]]
qnames = ["WWTCore", "csproj"]
version = "6.0.14.0"
age = 8

[[projects]]
qnames = ["WWTExplorer", "csproj"]
version = "6.0.907.0"
age = 0

[[projects]]
qnames = ["WorldWideTelescope", "csproj"]
version = "6.0.2.0"
age = 8

+++

+++ WWTExplorer3d/Layer.cs
@@ -178,10 +178,10 @@ namespace TerraViewer
         public virtual double[] GetParams()
         {
             double[] paramList = new double[5];
-            paramList[0] = color.R/255;
-            paramList[1] = color.G/255;
-            paramList[2] = color.B/255;
-            paramList[3] = color.A/255;
+            paramList[0] = ((double)color.R) / 255;
+            paramList[1] = ((double)color.G) / 255;
+            paramList[2] = ((double)color.B) / 255;
+            paramList[3] = ((double)color.A) / 255;
             paramList[4] = opacity;
